refac(ui): move install step of onboarding into a component
