+++ src/i18n/locales/en/translation.json
@@ -48,6 +48,7 @@
       "error": "Error loading Player Credentials",
       "form": {
         "button": "Submit",
+        "hint": "* Required Field",
         "password": {
           "description": "Stored using your Operating System's native credential manager.",
           "placeholder": "password",
@@ -61,8 +62,7 @@
           "description": "Stored using your Operating System's native credential manager.",
           "placeholder": "username",
           "title": "VU Username"
-        },
-        "hint": "* Required Field"
+        }
       },
       "loading": "Loading Users...",
       "sheet": {
@@ -227,12 +227,16 @@
     },
     "button": {
       "complete": "Skip / Complete Onboarding",
-      "continue": "Skip / Continue"
+      "continue": "Skip / Continue",
+      "refresh": "Refresh Check",
+      "verify": "Verify"
     },
     "error": "Failed to load Onboarding component",
     "failure": "Failed to complete onboarding",
     "header": "First Time Setup",
     "install": {
+      "complete": "Installation Complete!",
+      "description": "Download the latest version of Venice Unleashed",
       "dev": {
         "button": "Choose Install Location",
         "dialog": {
@@ -279,11 +283,19 @@
           },
           "installComplete": "VU Installation Completed"
         }
-      }
+      },
+      "success": "VU is ready to use.",
+      "title": "Download Venice Unleashed"
     },
     "loading": "Loading Onboarding component",
     "toast": {
-      "activateBF3": "Activating BF3"
+      "activateBF3": "Activating BF3",
+      "learnMore": "Learn more about Venice Unleashed"
+    },
+    "verify": {
+      "error": "Failed to verify installation status",
+      "installed": "Venice Unleashed is installed correctly!",
+      "notInstalled": "Venice Unleashed not found. Please install it to proceed."
     }
   },
   "servers": {
@@ -301,11 +313,11 @@
       },
       "loading": "Checking if Server Key Exists",
       "redoGuid": "Please re-enter your server GUID",
-      "title": "First Time Setup",
-      "subtitle": "Import Server Key & Setup Quick Join Here",
       "step1": "Download Server Key",
       "step2": "Copy Server GUID",
-      "step3": "Drag and Drop your key into this window"
+      "step3": "Drag and Drop your key into this window",
+      "subtitle": "Import Server Key & Setup Quick Join Here",
+      "title": "First Time Setup"
     },
     "loadouts": {
       "createLoadout": {
